pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astra"
version = "0.1.0"
description = "Hybrid local+cloud agentic voice assistant for Linux (Fedora)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Astra Maintainers"}]
license = {text = "MIT"}
keywords = ["assistant", "voice", "fastapi", "agent", "linux"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
]

# Runtime dependencies
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "cryptography>=43.0.0",
  "pyttsx3>=2.90",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
